+++ src/services/actions/cart-burger-costructor.js
@@ -1,0 +1,4 @@
+export const DELETE_ITEM = 'DELETE_ITEM';
+export const ADDED_ITEM = 'ADDED_ITEM';
+export const DELETE_ITEM_ON_INDEX = 'DELETE_ITEM_ON_INDEX';
+export const CHANGE_ORDER_INGREDIENT_IN_BURGER = 'CHANGE_ORDER_INGREDIENT_IN_BURGER';
